% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{equalCovs_Matrix}
\alias{equalCovs_Matrix}
\title{A more effective functions for testing the equality of two covariance matrices which are high-dimension}
\usage{
equalCovs_Matrix(sam1, sam2)
}
\arguments{
\item{sam1}{First sample, it must be array with structure size1*p, p is the dimension of data}

\item{sam2}{Second sample, it must be array with structure size2*p, p is the dimension of data}
}
\value{
test statistics and p-values
}
\description{
use Cpp to speed up the codes in R package 'equalCovs'
}
\examples{
\dontrun{
library(mvtnorm)
p<-700 # the dimension of multivariate

theta1<-2
theta2<-1
mat1<-diag(theta1,p-1)
mat2<-diag(theta1+theta1*theta2,p-1)
mat3<-diag(theta2,p-2)

mat1<-rbind(mat1,rep(0,p-1))
mat2<-rbind(mat2,rep(0,p-1))
mat3<-rbind(mat3,rep(0,p-2),rep(0,p-2))

mat1<-cbind(rep(0,p),mat1)
mat2<-cbind(rep(0,p),mat2)
mat3<-cbind(rep(0,p),rep(0,p),mat3)
sigma1<-mat1+t(mat1)+diag(1+theta1^2,p)
sigma2<-mat2+t(mat2)+mat3+t(mat3)+diag(1+theta1^2+theta2^2,p)

n1<-80
n2<-80
s1<-rmvnorm(n1,runif(p,0,5),sigma1) # generate the samples
s2<-rmvnorm(n2,runif(p,-3,3),sigma2)

test the result of three functions
library(equalCovs)
equalCovs(s1,s2,n1,n2)
equalCovs_C(s1,s2)
equalCovs_Matrix(s1, s2)

#test the time of three functions

library(microbenchmark)
t<-microbenchmark(Cpp=equalCovs_C(s1,s2),Fort=equalCovs(sa1,s2,n1,n2),Mat=equalCovs_Matrix(s1,s2))
summary(t)[,c(1,3,5,6)]
}
}
