% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
GibbsC(n, a, b)
}
\arguments{
\item{n}{x will obey B(n,y) when y is fixed}

\item{a}{y will obey Beta(x+a,n-x+b) when x is fixed}

\item{b}{y will obey Beta(x+a,n-x+b) when x is fixed}
}
\value{
a random sample matrix
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
n <- 10
a <- 5
b <- 6
Gibbs_C <- GibbsC(n,a,b)
}
}
