\name{StatComp21026-package}
\alias{StatComp21026-package}
\alias{StatComp21026}
\docType{package}
\title{
  A R package for Statistical Computing course.
}
\description{
  two more effective functions for testing the equality of two covariance matrices.
}

\author{
Author:Sun Wenming

Maintainer: Sun Wenming <swm123456@mail.ustc.edu.cn>
}

\keyword{ High-dimensional Covariance }

